feat: revise ShareDialog
