+++ src/components/Dialogs/ShareDialog/Content.tsx
@@ -1,6 +1,7 @@
 import { Dialog, ShareButtons, Translate } from '~/components'
 
 import { TextId } from '~/common/enums'
+import { toLocale } from '~/common/utils'
 
 import Copy from './Copy'
 import styles from './styles.css'
@@ -27,61 +28,63 @@ const ShareDialogContent: React.FC<ShareDialogContentProps> = ({
   headerTitle,
   description,
   footerButtons,
-}) => (
-  <>
-    <Dialog.Header
-      title={headerTitle || 'share'}
-      closeDialog={closeDialog}
-      closeTextId="close"
-      mode={headerTitle ? 'inner' : 'hidden'}
-    />
+}) => {
+  const url = new URL(shareLink)
+  const pathnames = url.pathname.split('/')
+  const showTranslation = toLocale(pathnames[1]) !== ''
+  if (showTranslation) {
+    description = (
+      <Translate
+        zh_hant="分享這篇文章的翻譯版本"
+        zh_hans="分享这篇文章的翻译版本"
+        en="Share this article in translated version"
+      />
+    )
+  }
+  return (
+    <>
+      <Dialog.Header
+        title={headerTitle || 'share'}
+        closeDialog={closeDialog}
+        leftButton={
+          <Dialog.Header.CloseButton closeDialog={closeDialog} textId="close" />
+        }
+      />
 
-    <Dialog.Content>
-      {description && (
-        <section className="description">
-          {description}
+      <Dialog.Content>
+        {description && (
+          <section className="description">
+            {description}
+
+            <style jsx>{styles}</style>
+          </section>
+        )}
+
+        <section className="socials-container">
+          <section className="left">
+            <ShareButtons.LINE title={shareTitle} link={shareLink} />
+            <ShareButtons.WhatsApp title={shareTitle} link={shareLink} />
+            <ShareButtons.Telegram title={shareTitle} link={shareLink} />
+            <ShareButtons.Douban title={shareTitle} link={shareLink} />
+            <Copy link={shareLink} />
+          </section>
+
+          <section className="right">
+            <ShareButtons.Twitter
+              title={shareTitle}
+              link={shareLink}
+              tags={shareTags}
+            />
+            <ShareButtons.Facebook title={shareTitle} link={shareLink} />
+            <ShareButtons.Weibo title={shareTitle} link={shareLink} />
+            <ShareButtons.Email title={shareTitle} link={shareLink} />
+          </section>
 
           <style jsx>{styles}</style>
         </section>
-      )}
-
-      <section className="socials-container">
-        <section className="left">
-          <ShareButtons.LINE title={shareTitle} link={shareLink} />
-          <ShareButtons.WhatsApp title={shareTitle} link={shareLink} />
-          <ShareButtons.Telegram title={shareTitle} link={shareLink} />
-          <ShareButtons.Douban title={shareTitle} link={shareLink} />
-        </section>
-
-        <section className="right">
-          <ShareButtons.Twitter
-            title={shareTitle}
-            link={shareLink}
-            tags={shareTags}
-          />
-          <ShareButtons.Facebook title={shareTitle} link={shareLink} />
-          <ShareButtons.Weibo title={shareTitle} link={shareLink} />
-          <ShareButtons.Email title={shareTitle} link={shareLink} />
-        </section>
-
-        <style jsx>{styles}</style>
-      </section>
-
-      <Copy link={shareLink} />
-    </Dialog.Content>
-
-    <Dialog.Footer>
-      {footerButtons || (
-        <Dialog.Footer.Button
-          bgColor="grey-lighter"
-          textColor="black"
-          onClick={closeDialog}
-        >
-          <Translate id="close" />
-        </Dialog.Footer.Button>
-      )}
-    </Dialog.Footer>
-  </>
-)
+      </Dialog.Content>
+    </>
+  )
+}
 
 export default ShareDialogContent
