+++ src/components/Dialogs/ShareDialog/styles.css
@@ -1,5 +1,7 @@
 .socials-container {
+  display: flex;
   padding-top: var(--spacing-x-tight);
+  padding-bottom: var(--spacing-xx-loose);
 }
 
 .left,
@@ -52,11 +54,11 @@
 }
 
 .description {
-  @mixin border-bottom-grey;
+  @mixin border-top-grey;
 
-  padding: var(--spacing-x-tight) var(--spacing-base) var(--spacing-loose);
+  padding: var(--spacing-base) var(--spacing-base) var(--spacing-x-tight);
   font-size: var(--font-size-md-s);
   line-height: 1.5rem;
-  color: var(--color-grey-dark);
+  color: var(--color-black);
   text-align: center;
 }
